rename skills to actions
