+++ examples/chatgpt/src/chatgpt_bridge/actions/cutout.py
