+++ examples/chatgpt/src/chatgpt_bridge/actions/erase.py
