+++ examples/chatgpt/src/chatgpt_bridge/actions/recolor.py
