+++ examples/chatgpt/src/chatgpt_bridge/actions/shadow.py
