+++ examples/chatgpt/src/chatgpt_bridge/actions/undo.py
